% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_parser.R
\name{bib2DT}
\alias{bib2DT}
\title{Bib file as data.table}
\usage{
bib2DT(file, to_sort = FALSE)
}
\arguments{
\item{file}{\code{.bib} file.}

\item{to_sort}{Include only author, title, year, and date.}
}
\description{
Bib file as data.table
}

