% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brace_closes_at.R
\name{braces_closes_at}
\alias{braces_closes_at}
\title{Brace closes at}
\usage{
braces_closes_at(tex_line, position_of_opening_brace)
}
\arguments{
\item{tex_line}{A single line.}

\item{position_of_opening_brace}{An integer giving the position of the opening brace in question.}
}
\value{
The positions of the closing brace matching the opening braces at \code{position_of_opening_brace}.
}
\description{
Where do braces close?
}
