% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_biber.R
\name{check_biber}
\alias{check_biber}
\title{Check biber}
\usage{
check_biber(path = ".", rstudio = FALSE)
}
\arguments{
\item{path}{The path containing the \code{blg} file, following successful compilation.}

\item{rstudio}{Use the RStudio API?}
}
\description{
Check biber
}
