% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consecutive_words.R
\name{check_consecutive_words}
\alias{check_consecutive_words}
\title{Check consecutive typeset words}
\usage{
check_consecutive_words(path = ".", latex_file = NULL, md5sum.ok = NULL)
}
\arguments{
\item{path}{Path containing the latex file.}
}
\description{
Check consecutive typeset words
}

