% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_escapes.R
\name{check_escapes}
\alias{check_escapes}
\title{Check escapes}
\usage{
check_escapes(filename, .report_error)
}
\arguments{
\item{filename}{File in which to report the error}

\item{.report_error}{How the errors should be reported.}
}
\value{
An error if unescaped dollar signs are present in \code{filename}. Otherwise, \code{NULL} invisibly.
}
\description{
Checks file for unescaped dollar signs. 
With these present, there is a risk of constructions like \code{We gave $10 to a million people at a cost of $10~million dollars.},
which is valid syntax, but incorrectly formatted. Accordingly, math-mode must be more assertively requested using \code{\(..\)}.
}

