% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_literal_citations.R
\name{check_literal_citations}
\alias{check_literal_citations}
\title{Check that citations are all using cites}
\usage{
check_literal_citations(filename, .report_error)
}
\arguments{
\item{filename}{TeX document}

\item{.report_error}{Function to report errors}
}
\description{
Check that citations are all using cites
}

