% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_literal_xrefs.R
\name{check_literal_xrefs}
\alias{check_literal_xrefs}
\title{Check for hard-coded cross-references}
\usage{
check_literal_xrefs(filename, .report_error)
}
\arguments{
\item{filename}{The TeX file to check}

\item{.report_error}{How errors should be reported.}
}
\value{
An error, or if none found, \code{NULL} invisibly.
}
\description{
Check for hard-coded cross-references
}

