% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_preamble.R
\name{check_preamble}
\alias{check_preamble}
\title{Check the preamble of a document}
\usage{
check_preamble(filename, .report_error, pre_release = FALSE,
  release = FALSE)
}
\arguments{
\item{filename}{.tex file to check for errors}

\item{.report_error}{How should errors be reported}

\item{pre_release}{See \code{\link{checkGrattanReport}}.}

\item{release}{See \code{\link{checkGrattanReport}}.}
}
\description{
Check the preamble of a document
}
