% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quote_marks.R
\name{check_quote_marks}
\alias{check_quote_marks}
\title{Check quote marks in TeX}
\usage{
check_quote_marks(filename, .report_error)
}
\arguments{
\item{filename}{LaTeX filename.}

\item{.report_error}{A function determining how errors will be reported.}
}
\description{
Checks whether a closing quote has been used at the start of a word.
}
