% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{check_spelling}
\alias{check_spelling}
\title{Spellchecker for Grattan reports}
\usage{
check_spelling(filename, ignore.lines = NULL, known.correct = NULL)
}
\arguments{
\item{filename}{Path to a LaTeX file to check.}

\item{ignore.lines}{Integer vector of lines to ignore (due to possibly spurious errors).}

\item{known.correct}{Character vector of words known to be correct (which will never be raised by this function).}
}
\value{
If the spell check fails, the line at which the first error was detected, with an error message. If the check suceeds, \code{NULL} invisibly.
}
\description{
Spellchecker for Grattan reports
}
\details{
Uses the \code{en_AU} hunspell dictionary.
}

