% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands_used.R
\name{commands_used}
\alias{commands_used}
\title{List all unique commands in a document}
\usage{
commands_used(tex_lines)
}
\arguments{
\item{tex_lines}{A LaTeX document as read from \code{readr::read_lines}.}
}
\value{
A character vector of unique commands used in \code{tex_lines}.
}
\description{
List all unique commands in a document
}
\examples{
commands_used(c("A \\\\abc{d}", "\\\\def{x}"))

}
