% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_spelled_words.R
\docType{data}
\name{correctly_spelled_words}
\alias{correctly_spelled_words}
\title{List of correctly spelled words}
\format{A character vector of words as perl-regex patterns to skip during the spell check.}
\usage{
correctly_spelled_words
}
\description{
List of correctly spelled words
}
\keyword{datasets}
