% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_optional_LaTeX_argument.R
\name{extract_optional_LaTeX_argument}
\alias{extract_optional_LaTeX_argument}
\title{Extract optional argument}
\usage{
extract_optional_LaTeX_argument(tex_lines, command_name, n = 1L,
  by.line = FALSE)
}
\arguments{
\item{tex_lines}{A character vector reading from a LaTeX document.}

\item{command_name}{Name of command (without backslash)}

\item{n}{Which optional argument to extract.}

\item{by.line}{Should the output be one row per command (\code{FALSE}, the default), 
with extracts concatenated via \code{paste0(..., collapse = "")}
or one row per line per command?}
}
\description{
Extract optional argument
}
