% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Extract authors from report}
\usage{
get_authors(filename)
}
\arguments{
\item{filename}{The filename containing the names of the authors.}
}
\value{
The names of Grattan staff who were authors in \code{filename}.
}
\description{
Extract authors from report
}

