% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isR_line_in_knitr.R
\name{isR_line_in_knitr}
\alias{isR_line_in_knitr}
\title{Is a line in knitr R or not?}
\usage{
isR_line_in_knitr(lines)
}
\arguments{
\item{lines}{Lines to check, as in the result of \code{readLines}. Not a filename.}
}
\value{
\code{TRUE} if in knitr chunk (including boundaries). \code{FALSE} otherwise.
}
\description{
Is a line in knitr R or not?
}
