% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_bib.R
\name{minimal_bib}
\alias{minimal_bib}
\title{Generate a minimal bibliography file}
\usage{
minimal_bib(path = ".", bbl.file = NULL, bib.files = NULL,
  out.bib = bib.files)
}
\arguments{
\item{path}{A directory containing a document after it has been run with \code{pdflatex}.}

\item{bbl.file}{A \code{.bbl} file.}

\item{bib.files}{The \code{.bib} file or files that were used by BibLaTeX to produce the bibliography.
If \code{NULL}, the default, the files are inferred from the contents of \code{\\\\addbibresource} within the (unique) \code{.tex} file are used.}

\item{out.bib}{The new file of bibliography.}
}
\description{
Generate a minimal bibliography file
}
