% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions_of_strings.R
\name{position_of_string}
\alias{position_of_string}
\alias{positions_of_all_strings}
\title{Position of strings}
\usage{
position_of_string(tex_line_split, command_split, end = TRUE)

positions_of_all_strings(tex_line, command_name, end = TRUE)
}
\arguments{
\item{tex_line_split}{A split line (via \code{strsplit(x, split = "")}).}

\item{command_split}{The string the position of which is desired, split (via \code{strsplit(x, split = "")}).}

\item{end}{(logical) Should the position of the \strong{end} of the string.
By default, \code{TRUE}; otherwise, the start of the string is chosen.}

\item{tex_line}{A line of text.}

\item{command_name}{The string the position of which is desired.}
}
\value{
The end (or start if end = FALSE) of the location of command
}
\description{
Position of strings
}
