% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tex_document.R
\name{read_tex_document}
\alias{read_tex_document}
\title{Read a LaTeX document}
\usage{
read_tex_document(file_root)
}
\arguments{
\item{file_root}{The root of the TeX file.}
}
\description{
Read a LaTeX document
}
