% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_LaTeX_argument.R
\name{replace_LaTeX_argument}
\alias{replace_LaTeX_argument}
\title{Replace contents of LaTeX argument}
\usage{
replace_LaTeX_argument(tex_lines, command_name, replacement,
  .dummy_replacer = "ZzZz")
}
\arguments{
\item{tex_lines}{Lines of text (as from \code{readLines}).}

\item{command_name}{Name of the command (without backslash) whose first
argument will be replaced.}

\item{replacement}{Replacement for the contents of each argument.}

\item{.dummy_replacer}{What to replace every \strong{letter} in the argument with during parsing.
Selecting a \code{.dummy_replacer} that is present in \code{tex_lines} is an error.}
}
\value{
A character vector with the first argument of \code{command_name} replaced with \code{replacement}.
}
\description{
Replace contents of LaTeX argument
}

