% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_arg_positions.R
\name{replace_nth_LaTeX_argument}
\alias{replace_nth_LaTeX_argument}
\title{Replace nth arguments}
\usage{
replace_nth_LaTeX_argument(tex_lines, command_name, fixed = TRUE, n = 1L,
  replacement = "correct", .dummy_replacement = "Zz")
}
\arguments{
\item{tex_lines}{A character vector of a LaTeX file read in from readLines.}

\item{command_name}{The command name without the initial backslash.}

\item{fixed}{Is \code{command_name} a pattern or a fixed expression?}

\item{n}{The number of mandatory arguments.}

\item{replacement}{What to replace the \code{n}th argument with.}

\item{.dummy_replacement}{An intermediate replacement value.
This value cannot be present in \code{tex_lines}.}
}
\description{
Replace nth arguments
}

