% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report2console}
\alias{report2gmail}
\alias{report2twitter}
\alias{report_error}
\title{Report errors to console and twitter}
\usage{
report2console(line_no = NULL, context = NULL, error_message = NULL,
  build_status = NULL, authors = NULL, report_name = NULL,
  extra_cat = NULL)

report2twitter(preamble = NULL, report_name, build_status, error_message,
  line_no = NULL, context = NULL, authors, extra_cat = NULL)

report2gmail(preamble = NULL, report_name, build_status, error_message,
  line_no = NULL, context = NULL, authors, extra_cat = NULL)
}
\arguments{
\item{line_no}{The line number locating the source of the error.}

\item{context}{THe content of the file to provide context to the error.}

\item{error_message}{The error message to display beyond the console.}

\item{build_status}{What should the build status be reported as?}

\item{authors}{Text to alert the authors (such as a twitter handle).}

\item{report_name}{Name of project whose errors are being reported.}

\item{extra_cat}{Character vector extra messages.}

\item{preamble}{Message to appear before other messages.}
}
\description{
Report errors to console and twitter
}

