% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_editorial_square_brackets.R
\name{rm_editorial_square_brackets}
\alias{rm_editorial_square_brackets}
\title{Remove editorial square brackets}
\usage{
rm_editorial_square_brackets(tex_lines)
}
\arguments{
\item{tex_lines}{Lines (as from \code{readLines}).}
}
\description{
Change text such as \code{phas[e] out} to \code{phase out},
without removing square brackets denoting optional arguments.
}
\examples{
x <- "the BCA's call to `urgently phas[e] out all side deals'"
rm_editorial_square_brackets(x)

}
