% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexCheckR-package.R
\name{TexCheckR-package}
\alias{TexCheckR-package}
\title{TexCheckR}
\description{
Checks LaTeX documents and .bib files for typing errors, such as spelling errors, incorrect quotation marks. Also provides useful functions for parsing and linting bibliography files.
}
