% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8ToLaTeX.R
\name{utf8ToLaTeX}
\alias{utf8ToLaTeX}
\title{Translate UTF-8 to ASCII with LaTeX}
\usage{
utf8ToLaTeX(input, filename = NULL, text = NULL, outfile = filename,
  rstudio = FALSE, verbose = FALSE)
}
\arguments{
\item{input}{The text file containing UTF-8 to read}

\item{filename, text}{May be provided instead of \code{input}.}

\item{outfile}{Where to direct the result. By default, the console unless \code{filename} is provided or \code{input} resolves to a filename, in which case the filename again (i.e. the filename is overwritten with the result).}

\item{rstudio}{In case of a problem during the conversion, use the RStudio API to pop to the location in the filename. Only applicable if \code{input} resolves to a file.}

\item{verbose}{Be chatty and report progress?}
}
\value{
The filename with new contents
}
\description{
Attempts to convert a UTF-8 file to one where all non-ASCII characters are replaced by control sequences which resolve in LaTeX or BibTeX to the equivalent printed characters. Inteded to be used when a bibliography needs to be fed to BibTeX (which does not support UTF-8).
}
