% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_English_contractions.R
\docType{data}
\name{valid_English_contractions}
\alias{valid_English_contractions}
\title{Valid English contractions}
\format{An object of class \code{character} of length 110.}
\source{
https://gist.githubusercontent.com/J3RN/ed7b420a6ea1d5bd6d06/raw/acda66b325a2b4d7282fb602a7551912cdc81e74/contractions.txt
}
\usage{
valid_English_contractions
}
\description{
List of words which should never raise a spelling error.
}
\keyword{datasets}
