% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veto_sic.R
\name{veto_sic}
\alias{veto_sic}
\title{Veto sic}
\usage{
veto_sic(tex_lines, quote = TRUE, sentence = !quote, words_ante = 1L)
}
\arguments{
\item{tex_lines}{A character vector.}

\item{quote}{(logical, default: \code{TRUE}) Veto words after the previous opening quote (\emph{i.e.} backtick) symbol.}

\item{sentence}{(logical, default: \code{TRUE}) Veto words before [sic] in the same sentence.
(The start of a sentence is taken to be the location of the capital letter which is preceded 
by white space and a full stop.)}

\item{words_ante}{The number of words to exclude. Ignored if \code{quote} or \code{sentence} is \code{TRUE}.}
}
\description{
Vetoes words in a LaTeX document that are marked '[sic]' 
for the purpose of spell checking by replacing them (and `[sic]` itself) with white space of 
equal length.
}
