% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weld_bmillion.R
\name{weld_bmillion}
\alias{weld_bmillion}
\title{Unbreaking spaces between billion and million}
\usage{
weld_bmillion(filename, outfile = filename)
}
\arguments{
\item{filename}{A LaTeX or knitr file.}
}
\value{
NULL. This function is called for its side-effect: rewriting \code{filename} with \code{30 million} changed to \code{30~million}.
}
\description{
Unbreaking spaces between billion and million
}

