% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrongly_spelled_words.R
\docType{data}
\name{wrongly_spelled_words}
\alias{wrongly_spelled_words}
\title{List of wrongly spelled words}
\format{A regex of patterns to raise as spelling errors.}
\usage{
wrongly_spelled_words
}
\description{
List of wrongly spelled words
}
\keyword{datasets}
