% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_bib_duplicates.R
\name{any_bib_duplicates}
\alias{any_bib_duplicates}
\title{Are any bib entries duplicated?}
\usage{
any_bib_duplicates(bib.files)
}
\arguments{
\item{bib.file}{File to check for duplicates}
}
\description{
Are any bib entries duplicated?
}
