% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_parser.R
\name{bib_parser}
\alias{bib2DT}
\alias{bib_parser}
\alias{reorder_bib}
\title{Functions for parsing .bib files}
\usage{
bib2DT(file.bib, to_sort = FALSE)

reorder_bib(file.bib, outfile.bib = file.bib)
}
\arguments{
\item{file.bib}{\code{.bib} file.}

\item{to_sort}{Include only author, title, year, and date.}

\item{outfile.bib}{File to write the reordered bib to. Defaults to \code{file.bib}.}
}
\description{
Functions for parsing .bib files
}
\details{
\code{bib2DT} returns a \code{data.table} of the entries in \code{file.bib}. The function
\code{reorder_bib} rewrites \code{file.bib}, to put it in surname, year, title, line number order.
}

