% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dashes.R
\name{check_dashes}
\alias{check_dashes}
\title{Check dashes entered as hyphens}
\usage{
check_dashes(filename)
}
\arguments{
\item{filename}{A tex or Rnw file.}
}
\value{
File stops and \code{cat()}s on any line where a hyphen is surrounded by a space.
}
\description{
Check dashes entered as hyphens
}

