% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footnote_typography.R
\name{check_footnote_typography}
\alias{check_footnote_typography}
\title{Check footnote typography}
\usage{
check_footnote_typography(filename, ignore.lines = NULL)
}
\arguments{
\item{filename}{A LaTeX file.}
}
\value{
Called for its side-effect.
}
\description{
Check footnote typography
}

