% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_footnote_typography.R
\name{check_footnote_typography}
\alias{check_footnote_typography}
\title{Check footnote typography}
\usage{
check_footnote_typography(filename, ignore.lines = NULL, .report_error)
}
\arguments{
\item{filename}{A LaTeX file.}

\item{ignore.lines}{Lines to ignore (for example, those using the word 'footnote').}

\item{.report_error}{A function to provide context to any errors.}
}
\value{
Called for its side-effect.
}
\description{
Check footnote typography
}
\details{
This function when applied to a LaTeX file will throw an error if: (1) footnotes
}
