% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labels.R
\name{check_labels}
\alias{check_labels}
\title{Check labels}
\usage{
check_labels(filename, .report_error)
}
\arguments{
\item{filename}{The LaTeX source file to check.}

\item{.report_error}{The function to provide context to the error.}
}
\value{
\code{NULL}, invisibly if labels check out. An error otherwise.
}
\description{
Check labels
}
\details{
Checks each label has a prefix and the prefix is one of the following:
\code{fig:},
\code{tbl:},
\code{box:},
\code{chap:},
\code{sec:},
\code{subsec:},
\code{subsubsec:},
\code{subsubsec:},
\code{para:}
\code{paragraph:}.
Checks also that chapter labels are marked with \code{chap:}.
}

