% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repetitive_xrefs.R
\name{check_repetitive_xrefs}
\alias{check_repetitive_xrefs}
\title{Check repetitive cross-references}
\usage{
check_repetitive_xrefs(filename)
}
\arguments{
\item{filename}{A LaTeX file}
}
\description{
Check repetitive cross-references
}

