% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_xrefs.R
\name{check_xrefs}
\alias{check_xrefs}
\title{Check cross-references}
\usage{
check_xrefs(filename, .report_error)
}
\arguments{
\item{filename}{A LaTeX file}

\item{.report_error}{The function to provide context to the error.}
}
\description{
Check cross-references that are repetitive or incorrect case.
}
