% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrattanReport.R
\name{checkGrattanReport}
\alias{checkGrattanReport}
\title{Check Grattan Report}
\usage{
checkGrattanReport(path = ".", output_method = c("console", "twitter",
  "gmailr"), compile = FALSE, final = FALSE, release = FALSE)
}
\arguments{
\item{path}{Path to search for the tex source file.}

\item{output_method}{How errors should be reported.}

\item{compile}{Should \code{pdflatex} be run on the report so the log be checked?}

\item{final}{Should the document be assumed to be final? Runs additional checks.}

\item{release}{Should a final pdf be prepared for publication?}
}
\value{
Called for its side-effect.
}
\description{
Check Grattan Report
}

