% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrattanReport.R
\name{checkGrattanReport}
\alias{checkGrattanReport}
\title{Check Grattan Report}
\usage{
checkGrattanReport(path = ".", output_method = c("console", "twitter",
  "gmailr"))
}
\arguments{
\item{path}{Path to search for the filename.}

\item{output_method}{How errors should be reported.}
}
\value{
Called for its side-effect.
}
\description{
Check Grattan Report
}

