% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrattanReport.R
\name{checkGrattanReport}
\alias{checkGrattanReport}
\title{Check Grattan Report}
\usage{
checkGrattanReport(path = ".", output_method = c("console", "twitter",
  "gmailr"), compile = FALSE, pre_release = FALSE, release = FALSE,
  .proceed_after_rerun, .no_log = FALSE, embed = TRUE,
  update_grattan.cls = TRUE)
}
\arguments{
\item{path}{Path to search for the tex source file.}

\item{output_method}{How errors should be reported.}

\item{compile}{Should \code{pdflatex} be run on the report so the logs may be checked?}

\item{pre_release}{Should the document be assumed to be final? Runs additional checks.}

\item{release}{Should a final pdf be prepared for publication?}

\item{.proceed_after_rerun}{On the occasions where infinitely many passes of \code{pdflatex} 
are required, include this to skip the error. Note that this will result in false cross-references 
or incompletely formatted bibliographies.}

\item{.no_log}{Make no entry in the log file on the check's outcome.}

\item{embed}{If \code{FALSE}, not attempt to embed the fonts using Ghostscript is attempted. Useful if Ghostscript cannot easily be installed.
Set to \code{TRUE} for debugging or repetitive use (as in benchmarking).}

\item{update_grattan.cls}{Download \code{grattan.cls} from \url{https://github.com/HughParsonage/grattex/blob/master/grattan.cls}? 
Set to \code{FALSE} when checking the \code{grattex} repo itself.}
}
\value{
Called for its side-effect.
}
\description{
Check Grattan reports written is LaTeX for typing errors, significant warnings, 
and inconsistent style.
}
