% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LaTeX_argument.R
\name{extract_LaTeX_argument}
\alias{extract_LaTeX_argument}
\title{Extract LaTeX command argument}
\usage{
extract_LaTeX_argument(tex_lines, command_name, n = 1L, optional = FALSE)
}
\arguments{
\item{tex_lines}{LaTeX text.}

\item{command_name}{Name of command without backslash \code{\\textbf} corresponds to \code{command_name = "textbf"}.}

\item{n}{Extract the nth argument.}

\item{optional}{Extract the optional argument, rather than the mandatory arguments.}
}
\value{
A \code{data.table}, each row corresponding to each line in \code{tex_lines}
and each instance of \code{command_name}. If multiple commands are on the same line,
subsequent instances of commands appear on subsequent rows.
\describe{
\item{\code{starts}}{Start position of the argument.}
\item{\code{stops}}{Stop position of the argument.}
\item{\code{extract}}{The argument extracted.}
\item{\code{line_no}}{The line number (with respect to \code{seq_along(tex_lines)}).}
}
}
\description{
Extract LaTeX command argument
}
