% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Extract authors from report}
\usage{
get_authors(filename, include_editors = TRUE)
}
\arguments{
\item{filename}{The filename whose preamble contains the names of the authors.}

\item{include_editors}{(logical) Should editorial members of staff be included?}
}
\value{
The names of Grattan staff who were authors in \code{filename}.
}
\description{
Extract authors from report
}
