% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_mandatory_LaTeX_argument.R
\name{locate_mandatory_LaTeX_argument}
\alias{locate_mandatory_LaTeX_argument}
\title{Locate contents of LaTeX commands}
\usage{
locate_mandatory_LaTeX_argument(tex_lines, command_name, n = 1L,
  parsed_doc = NULL)
}
\arguments{
\item{tex_lines}{A character vector of a LaTeX document, 
-- for example as obtained from \code{readLines("mydoc.tex")}.}

\item{command_name}{The command (without backslash) whose arguments' locations are desired.}

\item{n}{Integer vector: which argument(s) to locate. If \code{n = NA}, the n-th argument
positions \emph{for all n}.}

\item{parsed_doc}{The result of \code{parse_tex(tex_lines)}.}
}
\description{
Provides the locations of LaTeX commands with mandatory arguments.
}
