% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tex.R
\name{parse_tex}
\alias{parse_tex}
\title{Parse LaTeX lines}
\usage{
parse_tex(tex_lines, delim1 = "{", delim2 = "}")
}
\arguments{
\item{tex_lines}{Character vector (as read from a \code{.tex} file).}

\item{delim1, delim2}{Opening and closing characters for each group.}
}
\value{
A \code{data.table} where each row identifies a unique character in \code{tex_lines}.
\describe{
\item{\code{line_no}}{Matches the index of \code{tex_lines}.}
\item{\code{char_no}}{The character within \code{line_no}.}
\item{\code{char}}{The character. A single character.}
\item{\code{tex_group}}{The TeX group by default. Any delimiters can be used.}
}
}
\description{
Parse LaTeX lines
}
