% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report_error}
\alias{report2console}
\title{Report errors to console}
\usage{
report2console(file = NULL, line_no = NULL, column = NULL,
  context = NULL, error_message = NULL, advice = NULL,
  build_status = NULL, extra_cat_ante = NULL, extra_cat_post = NULL,
  caret = FALSE, rstudio = FALSE, log_file = NULL,
  log_file_sep = "|", silent = FALSE,
  halt = getOption("TeXCheckR.halt_on_error", FALSE),
  as_tbl = getOption("TeXCheckR.error_as_tbl", FALSE))
}
\arguments{
\item{file}{The file in which the error occurred.}

\item{line_no}{The line number locating the source of the error.}

\item{column}{The position on the line to identify the error (usually following the error).}

\item{context}{The content of the file, to provide context to the error.}

\item{error_message}{The error message to display beyond the console.}

\item{advice}{Advice to the user: how should the detected error be resolved in general?}

\item{build_status}{What should the build status be reported as?}

\item{extra_cat_ante}{Character vector extra messages (placed before \code{context}).}

\item{extra_cat_post}{Character vector extra messages (placed after \code{context}).}

\item{caret}{(logical, default: \code{FALSE}) Should a caret symbol be placed beneath the 
\code{context} to point to the location of the error? The caret will be inserted on a
new line after \code{error_message} and \code{extra_cat_post}.

Length-one integer values of \code{caret} are permitted and will be interpreted as the
number of caret symbols to be inserted at the position.}

\item{rstudio}{If available, should the report be allowed to modify the RStudio session (for example, to pop to the location of the error)?}

\item{log_file}{Optionally, path to a log file on which \code{error_message} will be written.}

\item{log_file_sep}{How should the log file's fields be separated? By default, with a pipe (as tabs are common within error messages).}

\item{silent}{(logical, default: \code{FALSE}) Suppress all output.}

\item{halt}{Should failures halt via \code{stop} or just display a message in the console?}

\item{as_tbl}{Return a list. Experimental.}
}
\description{
Report errors to console
}
