% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_error.R
\name{report_error}
\alias{report_error}
\alias{report2console}
\title{Report errors to console and twitter}
\usage{
report2console(line_no = NULL, context = NULL, error_message = NULL,
  advice = NULL, build_status = NULL, extra_cat_ante = NULL,
  extra_cat_post = NULL, log_file = NULL)
}
\arguments{
\item{line_no}{The line number locating the source of the error.}

\item{context}{THe content of the file to provide context to the error.}

\item{error_message}{The error message to display beyond the console.}

\item{advice}{Advice to the user: how should the detected error be resolved in general?}

\item{build_status}{What should the build status be reported as?}

\item{extra_cat_ante}{Character vector extra messages (placed before \code{context}).}

\item{extra_cat_post}{Character vector extra messages (placed after \code{context}).}

\item{log_file}{Optionally, path to a log file on which \code{error_message} will be written.}
}
\description{
Report errors to console and twitter
}
