% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_sentences.R
\name{separate_sentences}
\alias{separate_sentences}
\title{Put sentences on their own line}
\usage{
separate_sentences(filename, hanging_footnotes = FALSE)
}
\arguments{
\item{filename}{A tex or knitr file in which to separate sentences.}

\item{hanging_footnotes}{(logical, default: \code{FALSE}) Should footnotes be indented?}
}
\value{
NULL. The function is called for its side-effect: rewriting \code{filename} with separated sentences.
}
\description{
Put sentences on their own line
}
