% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexCheckR-package.R
\name{TeXCheckR-package}
\alias{TeXCheckR-package}
\title{TeXCheckR}
\description{
Checks LaTeX documents and \code{.bib} files for typing errors, such as spelling errors, incorrect quotation marks. Also provides useful functions for parsing and linting bibliography files.
}
