% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bibliography.R
\name{validate_bibliography}
\alias{validate_bibliography}
\title{Validate bibliography according Grattan style}
\usage{
validate_bibliography(path = ".", file = NULL, .report_error)
}
\arguments{
\item{path}{Containing the bib file.}

\item{file}{The bib file if specified.}

\item{.report_error}{How errors should be reported.}
}
\value{
\code{NULL} if bibliography validated.
}
\description{
Validate bibliography according Grattan style
}
